requirements.txt
fastapi
uvicorn
openai
pandas
matplotlib
python-multipart

fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
requests==2.31.0
pandas==2.1.3
numpy==1.24.3
matplotlib==3.8.2
plotly==5.17.0
seaborn==0.13.0
beautifulsoup4==4.12.2
lxml==4.9.3
openpyxl==3.1.2
pdfplumber==0.9.0
playwright==1.40.0
duckdb==0.9.2
paramiko==3.3.1
python-dotenv==1.0.0
Pillow==10.1.0



